validación vistas de registro
