+++ src/pages/registro.js
@@ -1,9 +1,10 @@
 import { signInGoogle } from '../lib/auth-google.js';
 import { signInEmail } from '../lib/auth-email.js';
+import { expresiones, validateField } from '../utils.js';
 
 export default {
-  path: '#registro',
-  template: `<div class='paws-image'>
+    path: '#registro',
+    template: `<div class='paws-image'>
   <img class='paws' src='images/animalsBackground.png' alt='paws' />
 </div>
   <div class='container-form'>
@@ -39,15 +40,37 @@ export default {
   </div>
   
             `,
-  script: () => {
-    const btn = document.querySelector('#btn-register');
-    btn.addEventListener('click', () => {
-      const email = document.querySelector('#email').value;
-      const password = document.querySelector('#password').value;
-      signInEmail(email, password);
-    });
+    script: () => {
+        const btn = document.querySelector('#btn-register');
+        const inputs = document.querySelectorAll('#form input');
 
-    const btnSignInGoogle = document.querySelector('#btnSignInGoogle');
-    btnSignInGoogle.addEventListener('click', signInGoogle);
-  },
+        btn.addEventListener('click', () => {
+            const email = document.querySelector('#email').value;
+            const password = document.querySelector('#password').value;
+            signInEmail(email, password);
+        });
+
+        const btnSignInGoogle = document.querySelector('#btnSignInGoogle');
+        btnSignInGoogle.addEventListener('click', signInGoogle);
+
+        const validarForm = (e) => {
+            switch (e.target.name) {
+                case 'name':
+                    validateField(expresiones.name, e.target, 'name');
+                    break;
+                case 'email':
+                    validateField(expresiones.email, e.target, 'email');
+                    break;
+                case 'password':
+                    validateField(expresiones.password, e.target, 'password');
+                    break;
+                default:
+            }
+        };
+
+        inputs.forEach((input) => {
+            input.addEventListener('keyup', validarForm);
+            input.addEventListener('blur', validarForm);
+        });
+    },
 };
